ajustes nos layouts para notificacoes toast entre login e logout
