+++ frontend/src/app/(default)/layout.tsx
@@ -1,50 +1,15 @@
-import { Geist, Geist_Mono } from "next/font/google";
 import { Navbar, Footer } from "@/app/components";
-import "@/styles/globals.css"
-import { Metadata, Viewport } from "next";
-import { AuthProvider } from "@/lib/context/AuthContext";
-import ToastProvider from "../components/toast-provider";
 
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
-
-export const metadata: Metadata = {
-  title: "Sanca Brechó",
-  description: "Compre e venda entre universitários de São Carlos",
-  icons: [{ rel: "icon", url: "/favicon.ico" }],
-};
-
-export const viewport: Viewport = {
-  width: 'device-width',
-  initialScale: 1.0,
-  maximumScale: 1.0,
-  userScalable: false,
-}
-
-export default function RootLayout({
+export default function DefaultLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
   return (
-    <html lang="pt-BR">
-      <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
-      >
-        <AuthProvider>
-          <Navbar />
-          <ToastProvider /> {/*Para notificacoes Toast*/}
-          {children}
-          <Footer />
-        </AuthProvider>
-      </body>
-    </html>
+    <>
+      <Navbar />
+      <main>{children}</main> 
+      <Footer />
+    </>
   );
 }
