+++ frontend/src/app/components/dynamic-toaster.tsx
@@ -1,0 +1,10 @@
+'use client'
+
+import dynamic from 'next/dynamic'
+
+const DynamicToaster = dynamic(
+    () => import('./toast-provider'),
+    { ssr: false }
+)
+
+export default DynamicToaster
