+++ frontend/src/app/layout.tsx
@@ -1,0 +1,35 @@
+// app/layout.tsx
+
+import { Geist, Geist_Mono } from "next/font/google";
+import "@/styles/globals.css";
+import { AuthProvider } from "@/lib/context/AuthContext";
+import DynamicToaster from "./components/dynamic-toaster";
+
+const geistSans = Geist({
+  variable: "--font-geist-sans",
+  subsets: ["latin"],
+});
+
+const geistMono = Geist_Mono({
+  variable: "--font-geist-mono",
+  subsets: ["latin"],
+});
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <html lang="pt-BR" suppressHydrationWarning={true}>
+      <body
+        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
+      >
+        <AuthProvider>
+          <DynamicToaster />
+          {children}
+        </AuthProvider>
+      </body>
+    </html>
+  );
+}
